backend/requirements.txt
# Flask e estensioni
Flask==3.0.0
Flask-CORS==4.0.0
Flask-Session==0.5.0

# Database
SQLAlchemy>=2.0.35
psycopg2-binary==2.9.9

# Utilità
python-dotenv==1.0.0
Werkzeug==3.0.1

# Per il futuro (download e conversione audio)
yt-dlp==2023.11.16
pydub==0.25.1
mutagen==1.47.0

# API esterne (per metadati)
requests==2.31.0
musicbrainzngs==0.7.1